Fix crash.
Fixed by not resetting the lcdTicks, lineClock and LY when the lcd setValue is
called for setting lcdc.
The issue was that a refresh() was going from pixel transfer mode
straight to oam mode without going through hblank (and thus not clearing
the activeObjs list). The immediatly following pixel transfer mode would call tranferPixels with a stale activeObjs list
which would then crash because it was drawing sprites that were more
than 16 distant from the scanline.